+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added Atlas Serializer support for `Guid`, `Uri`, `DateOnly`, and `TimeOnly`
+- Added TabInstance.CopyToClipboard(Call call...) variant to automatically show a Copied to Clipboard message
 
 ### Fixed
 - Fixed Atlas Serializer Decimal Member Handling with other Object Members